requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain_chroma
pandas
openpyxl
python-dotenv